[ghstack-poisoned]

+++ aten/src/ATen/native/quantized/cuda/FusedObsFakeQuant.cu
@@ -273,7 +273,7 @@ std::tuple<at::Tensor, at::Tensor> fused_moving_avg_obs_fake_quant_cuda(
     }
     _calculate_moving_average(
         y,
-        observer_on,
+        observer_on.to(at::kLong),
         running_min,
         running_max,
         averaging_const,
@@ -282,7 +282,7 @@ std::tuple<at::Tensor, at::Tensor> fused_moving_avg_obs_fake_quant_cuda(
   } else {
     _calculate_moving_average(
         x_contig,
-        observer_on,
+        observer_on.to(at::kLong),
         running_min,
         running_max,
         averaging_const,
@@ -295,7 +295,7 @@ std::tuple<at::Tensor, at::Tensor> fused_moving_avg_obs_fake_quant_cuda(
 
   _calc_moving_avg_qparams_helper(
       x_contig,
-      fake_quant_on,
+      fake_quant_on.to(at::kLong),
       running_min,
       running_max,
       scale_ptr,
@@ -316,7 +316,7 @@ std::tuple<at::Tensor, at::Tensor> fused_moving_avg_obs_fake_quant_cuda(
     }
   } else {
     return at::_fake_quantize_per_tensor_affine_cachemask_tensor_qparams(
-        x, scale, zero_point, fake_quant_on, qmin, qmax);
+        x, scale, zero_point, fake_quant_on.to(at::kLong), qmin, qmax);
   }
 }
 } // namespace at::native
